Rework the h-number implementation:
The original h-number implementation was definitely incorrect, it used counter indices as counter keys in incrementBranchCounts - runCoverage.getCovered() returns a list of non-zero indices, and counter.increment sets at a key. This was visible from logging + comparing the h_0 number to the total coverage number (they didn't match). Bad representation/implementation mismatch :/

This implementation eliminates the extra counter in favor of recomputing the values each time that they are requested. Abstractions are still not great, but this would be easier with the fast coverage implementation, anyway.